requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cryptography==36.0.2
cssselect2==0.5.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.13
django-import-export==2.8.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.11.3
lxml==4.8.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
pdfkit==0.6.1
pdfminer==20191125
Pillow==9.1.0
pycparser==2.21
pycryptodome==3.14.1
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyPDF2==1.27.4
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.0
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
xhtml2pdf==0.2.7
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
